# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.FriendGroupCreateRequest:
    properties:
      friend_id_list:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - name
    type: object
  model.FriendGroupUpdateRequest:
    properties:
      friend_id_list:
        description: 替换好友列表，允许空数组
        items:
          type: integer
        type: array
      group_id:
        type: integer
      name:
        description: 可选，修改分组名称
        type: string
    required:
    - group_id
    type: object
  model.FriendHandlerReq:
    properties:
      id:
        description: 记录id
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/model.Status'
        description: 状态
    type: object
...
  model.Response:
    description: 通用响应格式，用于包装所有 API 响应
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据部分
      message:
        description: 提示信息
        type: string
    type: object
...
  model.Status:
    enum:
    - 0
    - 1
    - 0
    - 1
    - 2
    type: integer
    x-enum-comments:
      Accept: 1 接受
      Disable: 0 禁用
      Enable: 1 启用
      Reject: 2 拒绝
      Todo: 0 待处理
    x-enum-descriptions:
    - 0 禁用
    - 1 启用
    - 0 待处理
    - 1 接受
    - 2 拒绝
    x-enum-varnames:
    - Disable
    - Enable
    - Todo
    - Accept
    - Reject
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  title: Swagger Example API
  version: "1.0"
paths:
  /friend/add:
    post:
      consumes:
      - application/json
      description: 当前用户向其他用户发送好友申请
      parameters:
      - description: 好友用户ID数组
        in: body
        name: data
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 添加好友申请
      tags:
      - Friend
  /friend/group_create:
    post:
      consumes:
      - application/json
      description: 给好友添加分组标签
      parameters:
      - description: 分组创建请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.FriendGroupCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 创建好友分组
      tags:
      - Friend
  /friend/group_delete:
    get:
      description: 根据分组 ID 删除好友分组
      parameters:
      - description: 分组ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除好友分组
      tags:
      - Friend
...
  /friend/group_update:
    post:
      consumes:
      - application/json
      description: 修改分组的名称等信息
      parameters:
      - description: 分组更新请求
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.FriendGroupUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 修改好友分组信息
      tags:
      - Friend
  /friend/handle_req:
    post:
      consumes:
      - application/json
      description: 同意或拒绝某条好友请求
      parameters:
      - description: 处理请求信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.FriendHandlerReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 同意或拒绝好友申请
      tags:
      - Friend
...
  /friend/remove:
    post:
      consumes:
      - application/json
      description: 删除当前用户的好友
      parameters:
      - description: 好友ID数组
        in: body
        name: data
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除好友
      tags:
      - Friend
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"